# openapi.yaml
openapi: 3.0.3
info:
  title: KYC Research API
  version: 1.0.0
  description: OpenAPI specification generated from backend/main.py endpoints /verify and /research
servers:
  - url: http://localhost:8000/amiki-kyc
paths:
  /verify:
    post:
      summary: Verify tenant
      operationId: verifyTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchRequest"
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/VerifyOk"
                  - $ref: "#/components/schemas/VerifyAmbiguous"
        '422':
          description: Validation error
  /research:
    post:
      summary: Run KYC research
      operationId: kycResearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchRequest"
      responses:
        '200':
          description: Research results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchSuccess"
        '422':
          description: Validation error
components:
  schemas:
    ResearchRequest:
      type: object
      required:
        - mieter
        - fragen
      properties:
        mieter:
          type: string
          description: Tenant name or identifier
        land:
          type: string
          nullable: true
        fragen:
          type: array
          items:
            type: string
          description: List of question IDs
        branche:
          type: string
          nullable: true
    TenantSuggestion:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        land:
          type: string
          nullable: true
        branche:
          type: string
          nullable: true
    VerifyOk:
      type: object
      properties:
        mieter:
          type: string
    VerifyAmbiguous:
      type: object
      properties:
        mieter_eingabe:
          type: string
        vorschlaege:
          type: array
          items:
            $ref: "#/components/schemas/TenantSuggestion"
    Antwort:
      type: object
      additionalProperties: true
    ResearchSuccess:
      type: object
      properties:
        mieter:
          type: string
        branche:
          type: string
          nullable: true
        land:
          type: string
          nullable: true
        antworten:
          type: array
          items:
            $ref: "#/components/schemas/Antwort"